refactor: fix name of isNumber to isNumeric
