+++ src/utils/addUnitIfNeeded.ts
@@ -1,5 +1,5 @@
 import { unitless } from './unitless'
-import { isNumeric } from './isNumber'
+import { isNumeric } from './isNumeric'
 
 // Taken from https://github.com/facebook/react/blob/b87aabdfe1b7461e7331abb3601d9e6bb27544bc/packages/react-dom/src/shared/dangerousStyleValue.js
 const addUnitIfNeeded = (property: string, value: any): string => {
