+++ src/utils/index.ts
@@ -3,7 +3,7 @@ export { addUnitIfNeeded } from './addUnitIfNeeded'
 export { createMediaQuery } from './createMediaQuery'
 export { defaultBreakpoints } from './defaultBreakpoints'
 export { getValue } from './getValue'
-export { isNumeric as isNumber } from './isNumber'
+export { isNumeric as isNumber } from './isNumeric'
 export { merge } from './merge'
 export { sort } from './sort'
 export { themeGet } from './themeGet'
